feat(scheduling): Enforce hourly caps for employee types
Added hourly limits for full-time (32-40 hrs/week) and part-time (12-20 hrs/week) employees.

+++ helper.py
@@ -9,4 +9,4 @@ def totalHours(fromTime, toTime):
     time2 = intoTime(toTime)
 
     hour_diff = time2 - time1
-    return hour_diff.total_seconds() / 3600
+    return int(hour_diff.total_seconds() / 3600)
